backend/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "chromadb-rag-backend"
version = "1.0.0"
description = "ChromaDB RAG API Backend"
requires-python = ">=3.11"
dependencies = [
    "fastapi==0.115.6",
    "uvicorn[standard]==0.34.0",
    "chromadb==0.4.24",
    "pypdf==3.17.0",
    "nltk==3.8.1",
    "python-dotenv==1.0.1",
    "aiofiles==23.2.1",
    "requests==2.31.0",
]
